# perf(python): optimise retrieval speed of values from `df.item` (~4-5x speedup)