We can get a ~4-5x speedup on `df.item()` by avoiding the unnecessary indirection that comes by passing-through to `__getitem__`. Instead, we can get away with a single `isinstance` check before calling the underlying `PySeries` directly.

(Another way to frame it would be that the _intrinsic_ speed of value retrieval hasn't changed at all, but the latency associated with setting up the call and returning the resulting value has been significantly reduced ;)

## Timings
Before/after:
```
1.52 µs ± 12.8 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops)
 338 ns ± 1.21 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops)
```