@alexander-beedie 2023-08-10
> You have a great eye for spotting improvements like this 😄 nice!

Thanks :) It's probably my past life as a C++ programmer trying to retain relevancy in the age of Python, haha...